functions/backup/requirements.txt
functions-framework==3.8.1
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-logging
